+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* Add `blikAliases` to LocalPaymentCompleted webhook and LocalPaymentDetails
+
 ## 3.36.0
 * Add `payerName`, `bic` and `iban_last_chars` for LocalPaymentCompleted webhook
 * Add `editPaypalVaultId` to `PayPalAccount`